Moved EditProfile under App component
