+++ frontend/src/app/profile-edit/profile-edit.graphql
@@ -1,0 +1,12 @@
+query MyProfile {
+    me {
+        email
+        firstName
+        lastName
+        gender
+        openToRecruiting
+        openToNewsletter
+        dateBirth
+        country
+    }
+}
